Make seller matching case-insensitive.
Hi! Thanks for making this extension!

Just a QoL proposal to make it easier when typing in seller names. I'd like to also propose a general punctuation cleaning regex, but at that point it's worth extracting `.trim().toLowerCase()` into a function, and I didn't want to go further than just a minor change, so I'll let this stand in for wider-spread potential text sanitization.

Thanks again!